Split out classes/interfaces from SqlConfigurableRetryLogicManager.cs
