+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Reliability/ISqlConfigurableRetryCommandSection.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Data.SqlClient
+{
+    internal interface ISqlConfigurableRetryCommandSection : ISqlConfigurableRetryConnectionSection
+    {
+        string AuthorizedSqlCondition { get; set; }
+    }
+}
