+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Reliability/SqlConfigurableRetryLogicManager.cs
@@ -86,25 +86,4 @@ namespace Microsoft.Data.SqlClient
             }
         }
     }
-
-    internal interface IAppContextSwitchOverridesSection
-    {
-        string Value { get; set; }
-    }
-
-    internal interface ISqlConfigurableRetryConnectionSection
-    {
-        TimeSpan DeltaTime { get; set; }
-        TimeSpan MaxTimeInterval { get; set; }
-        TimeSpan MinTimeInterval { get; set; }
-        int NumberOfTries { get; set; }
-        string RetryLogicType { get; set; }
-        string RetryMethod { get; set; }
-        string TransientErrors { get; set; }
-    }
-
-    internal interface ISqlConfigurableRetryCommandSection : ISqlConfigurableRetryConnectionSection
-    {
-        string AuthorizedSqlCondition { get; set; }
-    }
 }
